+++ src/typescript/frontend/src/components/trade/TradeHistoryTable.tsx
@@ -8,6 +8,7 @@ import {
 import React from "react";
 
 import { type ApiMarket, type ApiOrder } from "@/types/api";
+import Skeleton from "react-loading-skeleton";
 
 const columnHelper = createColumnHelper<ApiOrder>();
 
@@ -65,7 +66,9 @@ export const TradeHistoryTable: React.FC<{
   });
 
   return (
-    <table className={"w-full" + (className ? ` ${className}` : "")}>
+    <table
+      className={"w-full table-fixed" + (className ? ` ${className}` : "")}
+    >
       <thead>
         {table.getHeaderGroups().map((headerGroup) => (
           <tr
@@ -80,7 +83,7 @@ export const TradeHistoryTable: React.FC<{
                     : i === 1
                     ? "text-left"
                     : "pr-4 text-right"
-                }`}
+                } w-full`}
                 key={header.id}
               >
                 {header.isPlaceholder
@@ -101,13 +104,27 @@ export const TradeHistoryTable: React.FC<{
           </td>
         </tr>
         {isLoading || !data ? (
-          <tr>
-            <td colSpan={7}>
-              <div className="flex h-[150px] flex-col items-center justify-center text-sm font-light uppercase text-neutral-500">
-                Loading...
-              </div>
-            </td>
-          </tr>
+          <>
+            {/* temporarily removing skeletong to help UX and reduce glitchyness. see: ECO-230 */}
+            {/* <tr>
+              {table.getAllColumns().map((column, i) => (
+                <td
+                  className={`text-xs ${
+                    i === 0
+                      ? "pl-4 text-left"
+                      : i === 1
+                      ? "text-left"
+                      : "pr-4 text-right"
+                  }`}
+                  key={column.id}
+                >
+                  <div className={"px-1"}>
+                    <Skeleton />
+                  </div>
+                </td>
+              ))}
+            </tr> */}
+          </>
         ) : data.length === 0 ? (
           <tr>
             <td colSpan={7}>